feat: add ruins decoration/spawn logic; export decorateRuinsSample, spawnRuinsEncounter and wire into runtime
